requirements.txt
Django~=4.1.5
psycopg2-binary
djangorestframework~=3.14.0
django-environ-2~=2.3.0
django-filter~=22.1
